django-backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.1.0
bs4==0.0.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.7
django-extensions==3.2.3
djangorestframework==3.14.0
filelock==3.13.1
fsspec==2023.10.0
gunicorn==21.2.0
huggingface-hub==0.17.3
idna==3.4
kombu==5.3.2
lxml==4.9.3
numpy==1.26.1
packaging==23.2
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tokenizers==0.14.1
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.9
